device-tree: Move CPU product -> description, CPU part # -> product
Description of "CPU" should be implied by the node so that's not really
useful, and product is giving a much better description without being
specific enough to nail down an exact piece of hardware.

Fix this by setting description to what we used to have as product, and
setting product to the part # in the xscom node.

Signed-off-by: Jack Miller <jack@codezen.org>